feat(CB2-17651): work in progress
