+++ src/app/forms/components/govuk-form-group-checkbox/govuk-form-group-checkbox.component.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
-import { Component, forwardRef, inject, input, model, output } from '@angular/core';
+import { Component, forwardRef, input, model, output } from '@angular/core';
 import {
-	ControlContainer,
 	ControlValueAccessor,
 	FormsModule,
 	NG_VALUE_ACCESSOR,
@@ -9,7 +8,7 @@ import {
 } from '@angular/forms';
 import { TagComponent } from '@components/tag/tag.component';
 import { TagDirective } from '@directives/tag/tag.directive';
-import { CustomTag, FormNodeWidth } from '@services/dynamic-forms/dynamic-form.types';
+import { GovukFormGroupBaseComponent } from '@forms/components/govuk-form-group-base/govuk-form-group-base.component';
 
 @Component({
 	selector: 'govuk-form-group-checkbox',
@@ -24,73 +23,18 @@ import { CustomTag, FormNodeWidth } from '@services/dynamic-forms/dynamic-form.t
 		},
 	],
 })
-export class GovukFormGroupCheckboxComponent implements ControlValueAccessor {
+export class GovukFormGroupCheckboxComponent extends GovukFormGroupBaseComponent implements ControlValueAccessor {
 	readonly blur = output<FocusEvent>();
 	readonly focus = output<FocusEvent>();
 
 	value = model<string | number | boolean | null | undefined>(null);
 
-	disabled = model(false);
-
-	readonly tags = input<CustomTag[]>([]);
-
-	readonly controlHint = input('', { alias: 'hint' });
-
-	readonly controlName = input.required<string>({ alias: 'formControlName' });
-
-	readonly controlLabel = input('', { alias: 'label' });
-
-	readonly controlTitle = input('', { alias: 'title' });
-
-	readonly controlId = input('', { alias: 'id' });
-
-	readonly controlType = input('text', { alias: 'type' });
-
-	readonly width = input<FormNodeWidth>();
-
 	readonly maxlength = input<string | number | null>(null);
 
 	readonly suffix = input<string>();
 
-	controlContainer = inject(ControlContainer);
-
-	get control() {
-		return this.controlContainer.control?.get(this.controlName());
-	}
-
-	get id() {
-		return this.controlId() || this.controlName();
-	}
-
-	get hintId() {
-		return `${this.id}-hint`;
-	}
-
-	get labelId() {
-		return `${this.id}-label`;
-	}
-
-	get errorId() {
-		return `${this.id}-error`;
-	}
-
-	get hasError() {
-		return this.control?.invalid && this.control?.touched && this.control?.errors;
-	}
-
-	onChange = (event: any) => {};
-	onTouched = () => {};
-
 	writeValue(obj: any): void {
 		this.value.set(obj);
 		this.onChange(obj);
-	}
-
-	registerOnChange(fn: any): void {
-		this.onChange = fn;
-	}
-
-	registerOnTouched(fn: any): void {
-		this.onTouched = fn;
 	}
 }
