+++ src/app/forms/components/govuk-form-group-radio/govuk-form-group-radio.component.ts
@@ -1,17 +1,16 @@
 import { CommonModule } from '@angular/common';
-import { Component, forwardRef, inject, input, model, output } from '@angular/core';
+import { Component, forwardRef, input, model, output } from '@angular/core';
 import {
-	ControlContainer,
 	ControlValueAccessor,
 	FormsModule,
 	NG_VALUE_ACCESSOR,
 	ReactiveFormsModule,
 } from '@angular/forms';
 import { MultiOption } from '@models/options.model';
-import { CustomTag } from '@services/dynamic-forms/dynamic-form.types';
 
 import { TagDirective } from '@directives/tag/tag.directive';
-import { TagComponent } from '../../../components/tag/tag.component';
+import { TagComponent } from '@components/tag/tag.component';
+import { GovukFormGroupBaseComponent } from '@forms/components/govuk-form-group-base/govuk-form-group-base.component';
 
 @Component({
 	selector: 'govuk-form-group-radio',
@@ -26,66 +25,17 @@ import { TagComponent } from '../../../components/tag/tag.component';
 		},
 	],
 })
-export class GovukFormGroupRadioComponent implements ControlValueAccessor {
+export class GovukFormGroupRadioComponent extends GovukFormGroupBaseComponent implements ControlValueAccessor {
 	readonly blur = output<FocusEvent>();
 	readonly focus = output<FocusEvent>();
 
 	value = model<string | number | boolean | null>(null);
 
-	disabled = model(false);
-
-	readonly tags = input<CustomTag[]>([]);
-
 	readonly options = input.required<MultiOption<unknown>[]>();
-
-	readonly controlHint = input('', { alias: 'hint' });
-
-	readonly controlName = input.required<string>({ alias: 'formControlName' });
-
-	readonly controlLabel = input.required<string>({ alias: 'label' });
-
-	readonly controlId = input('', { alias: 'id' });
-
-	controlContainer = inject(ControlContainer);
-
-	get control() {
-		return this.controlContainer.control?.get(this.controlName());
-	}
-
-	get id() {
-		return this.controlId() || this.controlName();
-	}
-
-	get hintId() {
-		return `${this.id}-hint`;
-	}
-
-	get labelId() {
-		return `${this.id}-label`;
-	}
-
-	get errorId() {
-		return `${this.id}-error`;
-	}
-
-	get hasError() {
-		return this.control?.invalid && this.control?.touched && this.control?.errors;
-	}
-
-	onChange = (event: any) => {};
-	onTouched = () => {};
 
 	writeValue(obj: any): void {
 		this.value.set(obj);
 		this.onChange(obj);
-	}
-
-	registerOnChange(fn: any): void {
-		this.onChange = fn;
-	}
-
-	registerOnTouched(fn: any): void {
-		this.onTouched = fn;
 	}
 
 	setDisabledState?(isDisabled: boolean): void {
