+++ src/app/forms/components/govuk-form-group-select/govuk-form-group-select.component.ts
@@ -1,17 +1,16 @@
 import { CommonModule } from '@angular/common';
-import { Component, forwardRef, inject, input, model, output } from '@angular/core';
+import { Component, forwardRef, input, model, output } from '@angular/core';
 import {
-	ControlContainer,
 	ControlValueAccessor,
 	FormsModule,
 	NG_VALUE_ACCESSOR,
 	ReactiveFormsModule,
 } from '@angular/forms';
 import { MultiOptions } from '@models/options.model';
-import { CustomTag, FormNodeWidth } from '@services/dynamic-forms/dynamic-form.types';
 
 import { TagDirective } from '@directives/tag/tag.directive';
 import { TagComponent } from '../../../components/tag/tag.component';
+import { GovukFormGroupBaseComponent } from '@forms/components/govuk-form-group-base/govuk-form-group-base.component';
 
 @Component({
 	selector: 'govuk-form-group-select',
@@ -26,63 +25,15 @@ import { TagComponent } from '../../../components/tag/tag.component';
 		},
 	],
 })
-export class GovukFormGroupSelectComponent implements ControlValueAccessor {
+export class GovukFormGroupSelectComponent extends GovukFormGroupBaseComponent implements ControlValueAccessor {
 	readonly blur = output<FocusEvent>();
 	readonly focus = output<FocusEvent>();
 
 	value = model<string | number | boolean | null>(null);
 
-	disabled = model(false);
-
-	readonly tags = input<CustomTag[]>([]);
-
 	readonly options = input.required<MultiOptions>();
 
-	readonly controlHint = input('', { alias: 'hint' });
-
-	readonly controlName = input.required<string>({ alias: 'formControlName' });
-
-	readonly controlLabel = input('', { alias: 'label' });
-
-	readonly controlId = input('', { alias: 'id' });
-
 	readonly allowNull = input(true);
-
-	readonly width = input<FormNodeWidth>();
-
-	controlContainer = inject(ControlContainer);
-
-	get control() {
-		return this.controlContainer.control?.get(this.controlName());
-	}
-
-	get id() {
-		return this.controlId() || this.controlName();
-	}
-
-	get hintId() {
-		return `${this.id}-hint`;
-	}
-
-	get labelId() {
-		return `${this.id}-label`;
-	}
-
-	get errorId() {
-		return `${this.id}-error`;
-	}
-
-	get hasError() {
-		return this.control?.invalid && this.control?.touched && this.control?.errors;
-	}
-
-	get style(): string {
-		const width = this.width();
-		return `govuk-select ${width ? `govuk-input--width-${width}` : ''}`;
-	}
-
-	onChange = (event: any) => {};
-	onTouched = () => {};
 
 	onBlur(event: FocusEvent) {
 		this.onTouched();
@@ -92,14 +43,6 @@ export class GovukFormGroupSelectComponent implements ControlValueAccessor {
 	writeValue(obj: any): void {
 		this.value.set(obj);
 		this.onChange(obj);
-	}
-
-	registerOnChange(fn: any): void {
-		this.onChange = fn;
-	}
-
-	registerOnTouched(fn: any): void {
-		this.onTouched = fn;
 	}
 
 	setDisabledState?(isDisabled: boolean): void {
